+++ docs/js/cropper.js
@@ -5,7 +5,7 @@
  * Copyright 2015-present CIT Open Region
  * Released under the MIT license
  *
- * Date: 2021-06-16T05:29:34.056Z
+ * Date: 2021-06-16T10:37:42.072Z
  */
 
 (function (global, factory) {
@@ -246,6 +246,8 @@
     fitCanvasOnRotate: false,
     // Enable to scale the image
     scalable: true,
+    // Enable to skew the image
+    skewable: true,
     // Enable to zoom the image
     zoomable: true,
     // Enable to zoom the image by dragging touch
@@ -773,6 +775,8 @@
     var rotate = _ref.rotate,
         scaleX = _ref.scaleX,
         scaleY = _ref.scaleY,
+        skewX = _ref.skewX,
+        skewY = _ref.skewY,
         translateX = _ref.translateX,
         translateY = _ref.translateY;
     var values = [];
@@ -796,6 +800,14 @@
 
     if (isNumber(scaleY) && scaleY !== 1) {
       values.push("scaleY(".concat(scaleY, ")"));
+    }
+
+    if (isNumber(skewX) && skewX !== 0) {
+      values.push("skewX(".concat(skewX, "deg)"));
+    }
+
+    if (isNumber(skewY) && skewY !== 0) {
+      values.push("skewY(".concat(skewY, "deg)"));
     }
 
     var transform = values.length ? values.join(' ') : 'none';
@@ -961,7 +973,11 @@
         _ref6$scaleX = _ref6.scaleX,
         scaleX = _ref6$scaleX === void 0 ? 1 : _ref6$scaleX,
         _ref6$scaleY = _ref6.scaleY,
-        scaleY = _ref6$scaleY === void 0 ? 1 : _ref6$scaleY;
+        scaleY = _ref6$scaleY === void 0 ? 1 : _ref6$scaleY,
+        _ref6$skewX = _ref6.skewX,
+        skewX = _ref6$skewX === void 0 ? 0 : _ref6$skewX,
+        _ref6$skewY = _ref6.skewY,
+        skewY = _ref6$skewY === void 0 ? 0 : _ref6$skewY;
     var aspectRatio = _ref7.aspectRatio,
         naturalWidth = _ref7.naturalWidth,
         naturalHeight = _ref7.naturalHeight;
@@ -1016,6 +1032,7 @@
     context.translate(width / 2, height / 2);
     context.rotate(rotate * Math.PI / 180);
     context.scale(scaleX, scaleY);
+    context.transform(1, skewY * (Math.PI / 180), skewX * (Math.PI / 180), 1, 0, 0);
     context.imageSmoothingEnabled = imageSmoothingEnabled;
     context.imageSmoothingQuality = imageSmoothingQuality;
     context.drawImage.apply(context, [image].concat(_toConsumableArray(params.map(function (param) {
@@ -1388,9 +1405,15 @@
           imageData = this.imageData;
 
       if (transformed) {
+        var skewXMultiplier = imageData.skewX < 0 ? -1 : 1;
+        var skewYMultiplier = imageData.skewY < 0 ? -1 : 1;
+        var skewMultiplier = skewXMultiplier * skewYMultiplier < 0 ? -1 : 1;
+        var skewYIncrement = imageData.skewY ? Math.abs(imageData.naturalWidth * Math.abs(imageData.scaleX || 1) * Math.sin(imageData.skewY * (Math.PI / 180)) / Math.sin((90 - imageData.skewY) * (Math.PI / 180))) : 0;
+        var skewXIncrement = imageData.skewX ? Math.abs((imageData.naturalHeight * Math.abs(imageData.scaleY || 1) + skewYIncrement * skewMultiplier) * Math.sin(imageData.skewX * (Math.PI / 180)) / Math.sin((90 - imageData.skewX) * (Math.PI / 180))) : 0;
+
         var _getRotatedSizes = getRotatedSizes({
-          width: imageData.naturalWidth * Math.abs(imageData.scaleX || 1),
-          height: imageData.naturalHeight * Math.abs(imageData.scaleY || 1),
+          width: imageData.naturalWidth * Math.abs(imageData.scaleX || 1) + skewXIncrement,
+          height: imageData.naturalHeight * Math.abs(imageData.scaleY || 1) + skewYIncrement,
           degree: imageData.rotate || 0
         }),
             naturalWidth = _getRotatedSizes.width,
@@ -2811,6 +2834,58 @@
     },
 
     /**
+     * Skew the image on the x-axis.
+     * @param {number} skewX - The skew ratio on the x-axis.
+     * @returns {Cropper} this
+     */
+    skewX: function skewX(_skewX) {
+      var skewY = this.imageData.skewY;
+      return this.skew(_skewX, isNumber(skewY) ? skewY : 0);
+    },
+
+    /**
+     * Skew the image on the y-axis.
+     * @param {number} skewY - The skew ratio on the y-axis.
+     * @returns {Cropper} this
+     */
+    skewY: function skewY(_skewY) {
+      var skewX = this.imageData.skewX;
+      return this.skew(isNumber(skewX) ? skewX : 0, _skewY);
+    },
+
+    /**
+     * Skew the image
+     * @param {number} skewX - The skew ratio on the x-axis.
+     * @param {number} [skewY=skewX] - The skew ratio on the y-axis.
+     * @returns {Cropper} this
+     */
+    skew: function skew(skewX) {
+      var skewY = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : skewX;
+      var imageData = this.imageData;
+      var transformed = false;
+      skewX = Number(skewX);
+      skewY = Number(skewY);
+
+      if (this.ready && !this.disabled && this.options.skewable) {
+        if (isNumber(skewX)) {
+          imageData.skewX = skewX;
+          transformed = true;
+        }
+
+        if (isNumber(skewY)) {
+          imageData.skewY = skewY;
+          transformed = true;
+        }
+
+        if (transformed) {
+          this.renderCanvas(true, true);
+        }
+      }
+
+      return this;
+    },
+
+    /**
      * Get Cropper options
      * @returns {Object} Options
      */
@@ -2871,6 +2946,11 @@
         data.scaleY = imageData.scaleY || 1;
       }
 
+      if (options.scalable) {
+        data.skewX = imageData.skewX || 0;
+        data.skewY = imageData.skewY || 0;
+      }
+
       return data;
     },
 
@@ -2903,6 +2983,18 @@
 
           if (isNumber(data.scaleY) && data.scaleY !== imageData.scaleY) {
             imageData.scaleY = data.scaleY;
+            transformed = true;
+          }
+        }
+
+        if (options.scalable) {
+          if (isNumber(data.skewX) && data.skewX !== imageData.skewX) {
+            imageData.skewX = data.skewX;
+            transformed = true;
+          }
+
+          if (isNumber(data.skewY) && data.skewY !== imageData.skewY) {
+            imageData.skewY = data.skewY;
             transformed = true;
           }
         }
@@ -3402,6 +3494,8 @@
         var rotate = 0;
         var scaleX = 1;
         var scaleY = 1;
+        var skewX = 0;
+        var skewY = 0;
 
         if (orientation > 1) {
           // Generate a new URL which has the default orientation value
@@ -3421,6 +3515,11 @@
         if (options.scalable) {
           imageData.scaleX = scaleX;
           imageData.scaleY = scaleY;
+        }
+
+        if (options.skewable) {
+          imageData.skewX = skewX;
+          imageData.skewY = skewY;
         }
 
         this.clone();