+++ src/js/methods.js
@@ -392,6 +392,60 @@ export default {
   },
 
   /**
+   * Skew the image on the x-axis.
+   * @param {number} skewX - The skew ratio on the x-axis.
+   * @returns {Cropper} this
+   */
+  skewX(skewX) {
+    const { skewY } = this.imageData;
+
+    return this.skew(skewX, isNumber(skewY) ? skewY : 0);
+  },
+
+  /**
+   * Skew the image on the y-axis.
+   * @param {number} skewY - The skew ratio on the y-axis.
+   * @returns {Cropper} this
+   */
+  skewY(skewY) {
+    const { skewX } = this.imageData;
+
+    return this.skew(isNumber(skewX) ? skewX : 0, skewY);
+  },
+
+  /**
+   * Skew the image
+   * @param {number} skewX - The skew ratio on the x-axis.
+   * @param {number} [skewY=skewX] - The skew ratio on the y-axis.
+   * @returns {Cropper} this
+   */
+  skew(skewX, skewY = skewX) {
+    const { imageData } = this;
+    let transformed = false;
+
+    skewX = Number(skewX);
+    skewY = Number(skewY);
+
+    if (this.ready && !this.disabled && this.options.skewable) {
+      if (isNumber(skewX)) {
+        imageData.skewX = skewX;
+        transformed = true;
+      }
+
+      if (isNumber(skewY)) {
+        imageData.skewY = skewY;
+        transformed = true;
+      }
+
+      if (transformed) {
+        this.renderCanvas(true, true);
+      }
+    }
+
+    return this;
+  },
+
+  /**
    * Get Cropper options
    * @returns {Object} Options
    */
@@ -456,6 +510,11 @@ export default {
       data.scaleY = imageData.scaleY || 1;
     }
 
+    if (options.scalable) {
+      data.skewX = imageData.skewX || 0;
+      data.skewY = imageData.skewY || 0;
+    }
+
     return data;
   },
 
@@ -486,6 +545,18 @@ export default {
 
         if (isNumber(data.scaleY) && data.scaleY !== imageData.scaleY) {
           imageData.scaleY = data.scaleY;
+          transformed = true;
+        }
+      }
+
+      if (options.scalable) {
+        if (isNumber(data.skewX) && data.skewX !== imageData.skewX) {
+          imageData.skewX = data.skewX;
+          transformed = true;
+        }
+
+        if (isNumber(data.skewY) && data.skewY !== imageData.skewY) {
+          imageData.skewY = data.skewY;
           transformed = true;
         }
       }