+++ test/specs/events/crop.spec.js
@@ -14,7 +14,9 @@ describe('crop (event)', () => {
     const image = window.createImage();
 
     image.addEventListener('crop', (event) => {
-      expect(event.detail).to.be.an('object').that.has.all.keys(['x', 'y', 'width', 'height', 'rotate', 'scaleX', 'scaleY']);
+      expect(event.detail).to.be.an('object').that.has.all.keys(
+        ['x', 'y', 'width', 'height', 'rotate', 'scaleX', 'scaleY', 'skewX', 'skewY'],
+      );
       expect(event.detail.x).to.be.a('number');
       expect(event.detail.y).to.be.a('number');
       expect(event.detail.width).to.be.a('number');